requirements.txt
Authlib==1.3.2
google_api_python_client==2.142.0
google_auth_oauthlib==1.2.1
langchain_community==0.2.16
langchain_core==0.2.38
langchain_google_genai==1.0.10
protobuf==5.28.0
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
sentence_transformers==2.2.2
streamlit==1.38.0
torch==2.4.1

